requirement.txt

numpy
pandas
Flask
mysql-connector-python
pymysql
python-dotenv
scikit-learn